More refactoring, document new project structure.
