+++ dialogue.html
@@ -7,6 +7,6 @@
 </head>
 <body>
     <div id="dialogue-app"></div>
-    <script type="module" src="/src/devtools/dialogue.tsx"></script>
+    <script type="module" src="/src/devtools/DialogueGraphVisualizer.tsx"></script>
 </body>
 </html>
