+++ src/assets/3d/models/player_ref.fbx
