+++ src/core/battle/engine/battle.config.ts
@@ -1,7 +1,7 @@
-import ai_plain_icon from '@/assets/icons/battle-alerts/plain.png'
-import ai_heal_icon from '@/assets/icons/battle-alerts/heal.png'
-import ai_focus_icon from '@/assets/icons/battle-alerts/focus.png'
-import ai_mania_icon from '@/assets/icons/battle-alerts/mania.png'
+import ai_plain_icon from '@/features/battle/assets/icons/battle-alerts/plain.png'
+import ai_heal_icon from '@/features/battle/assets/icons/battle-alerts/heal.png'
+import ai_focus_icon from '@/features/battle/assets/icons/battle-alerts/focus.png'
+import ai_mania_icon from '@/features/battle/assets/icons/battle-alerts/mania.png'
 
 /** Delay between showing notifications/mults and dealing damage */
 export const PREANIM_MIN_DURATION = 750;
