+++ src/features/battle/assets/icons/battle-alerts/blank.png
