+++ src/features/battle/assets/icons/battle-alerts/mania.png
