+++ src/features/battle/assets/icons/battle-alerts/plain.png
