+++ src/features/battle/assets/icons/statuses/mania.png
