+++ src/features/battle/assets/icons/statuses/prep.png
