+++ src/features/battle/assets/icons/statuses/vuln.png
