+++ src/features/battle/assets/overlay-animations/shield/shield_opp1.png
