+++ src/features/battle/assets/overlay-animations/slash/slash_majes-min.png
