+++ src/features/battle/assets/overlay-animations/slash/slash_purpose.png
