+++ src/features/battle/assets/overlay-animations/slash/slashnorm.png
