+++ src/scenes/AnotherScene/AnotherScene.tsx
@@ -1,6 +1,6 @@
-import mapobj from './models/map.obj?url'
-import mapmtl from './models/map.mtl?url'
-import player_ref from '../shared_models/player_ref.fbx?url'
+import mapobj from './models/map.obj'
+import mapmtl from './models/map.mtl'
+import player_ref from '@/assets/3d/models/player_ref.fbx'
 import { createSignal, onMount } from 'solid-js'
 import Interactable from '@/3d/components/Interactable'
 import { Scene } from 'lume'
